# output/architecture/openapi.yaml
openapi: 3.1.2
...
info:
  title: Checkout sandbox API
  version: 1.0.0
  summary: Contrato HTTP del checkout invitado asíncrono
  description: |
    Contrato contract-first de Etapa 3 consumido por la construcción de Etapa 5.
    El servidor es autoridad de precio, moneda, stock, cotización y estado.
    Los datos completos de tarjeta nunca forman parte de esta API. Los valores
    efímeros de pago y aceptación sólo aparecen como campos de escritura sin
    ejemplos, defaults ni persistencia.
  x-status: STAGE_5_LOCAL_FAKE_VERIFIED
  x-source-baseline:
    requirements: 33 RF identities and 28 RNF identities
    acceptanceCriteria: 45
    scenarios: 51
    errors: 24
    dataFields: 72
    uat: 48
  x-design-decisions:
    oasVersion: OpenAPI 3.1.2 for reversible toolchain interoperability
    customerRoute: One nested PUT route under the checkout; no duplicate top-level write route
    stockRoute: Kept as a separate nominal read resource for the challenge rubric
    webhookRoute: One singular optional P1 route
    capability: HttpOnly cookie; raw value never enters JSON, URLs, logs or examples
    paymentSubmission: Every locally accepted payment submission and its semantic replay return the same 202 body and Location
    asyncOutcomes: Provider uncertainty and later incidents are represented by the local transaction resource
    localStateVocabulary: E0-E2 values NOT_SENT and ACKNOWLEDGED are preserved
...
servers:
  - url: https://checkout.example.invalid
...
security: []
...
paths:
  /api/v1/products:
    get:
      operationId: listProducts
      tags: [Products]
      summary: Listar productos activos y disponibilidad pública
      description: Consulta pública sin efectos de negocio ni contadores internos.
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      x-api-id: API-01
      x-status: IMPLEMENTED_LOCAL
      x-trace:
        requirements: [RF-01, RF-25, RF-16]
        acceptanceCriteria: [AC-US-01-01, AC-US-01-02]
        scenarios: [SC-US-01-01, SC-US-01-02]
        uat: [UAT-18, UAT-38]
        data: [DAT-01, DAT-02, DAT-03, DAT-04, DAT-05, DAT-06, DAT-07, DAT-10]
        errors: [ERR-02, ERR-20]
        uxFlows: [UXF-01, UXF-02]
      responses:
        '200':
          description: Colección pública de productos activos
          headers:
            Cache-Control: { $ref: '#/components/headers/RevalidateCache' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductCollectionResponse' }
              examples:
                availableProduct:
                  value:
                    items:
                      - productId: PRODUCT_ID_OPAQUE
                        sku: SKU_SYNTHETIC
                        name: PRODUCT_AVAILABLE
                        description: Producto sintético para la demostración.
                        imageUrl: https://assets.example.invalid/product.svg
                        unitPrice: { amountInCents: 2500000, currency: COP }
                        available: 1
                    count: 1
        '403': { $ref: '#/components/responses/OriginForbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/products/{productId}:
    get:
      operationId: getProduct
      tags: [Products]
      summary: Obtener un producto activo
      description: Consulta pública sin efectos; un ID nunca concede acceso privado.
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/CorrelationId'
      x-api-id: API-02
      x-status: IMPLEMENTED_LOCAL
      x-trace:
        requirements: [RF-01, RF-25]
        acceptanceCriteria: [AC-US-01-01, AC-US-01-03]
        scenarios: [SC-US-01-01, SC-US-01-03]
        uat: [UAT-37]
        data: [DAT-01, DAT-02, DAT-03, DAT-04, DAT-05, DAT-06, DAT-07, DAT-10]
        errors: [ERR-02, ERR-04, ERR-20]
        uxFlows: [UXF-01, UXF-02]
      responses:
        '200':
          description: Producto público
          headers:
            Cache-Control: { $ref: '#/components/headers/RevalidateCache' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductResponse' }
        '403': { $ref: '#/components/responses/OriginForbidden' }
        '404': { $ref: '#/components/responses/ProductNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    CheckoutCapability:
      type: apiKey
      in: cookie
      name: __Secure-checkout_cap
      description: |
        Capability aleatoria de al menos 256 bits emitida sólo mediante una
        cookie HttpOnly, Secure, SameSite=Strict y Path=/api/v1. El servidor
        conserva únicamente su HMAC/hash. Ausencia, expiración, alteración o
        relación ajena producen el mismo 404. No hay valores ni ejemplos.
      x-data-class: C3
      x-raw-storage: COOKIE_AND_EPHEMERAL_GUARD_MEMORY_ONLY
    ProviderWebhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: |
        Contrato abstracto P1 para una firma de evento. El nombre y algoritmo
        del adapter real permanecen bloqueados hasta SPK-02; nunca se registra
        ni se ejemplifica el valor.
      x-data-class: C3
      x-status: BLOCKED_REAL_PROVIDER
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      description: Identificador público opaco; no concede autorización.
      schema: { $ref: '#/components/schemas/OpaqueId' }
      example: PRODUCT_ID_OPAQUE
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Hint opaco no confiable; el servidor nunca lo refleja ni registra y siempre genera un ID seguro nuevo.
      schema: { $ref: '#/components/schemas/CorrelationId' }
  headers:
    CorrelationId:
      description: Identificador opaco de la request; puede cambiar entre replays.
      schema: { $ref: '#/components/schemas/CorrelationId' }
      example: CORRELATION_ID_SAFE
...
    NoStore:
      description: Prohíbe almacenamiento de la respuesta.
      schema:
        type: string
        const: no-store
      example: no-store
    RevalidateCache:
      description: La disponibilidad se revalida y nunca autoriza una reserva.
      schema:
        type: string
        const: no-cache
      example: no-cache
...
  responses:
...
    OriginForbidden:
      description: ERR-02 ORIGIN_FORBIDDEN; no revela recursos.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ProductNotFound:
      description: ERR-04 PRODUCT_NOT_FOUND
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    InternalError:
      description: |
        ERR-20 antes de I/O externo ambiguo. Si pudo existir envío, el recurso
        aceptado permanece consultable y se proyecta como ERR-14, no como 500.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    OpaqueId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: '^[A-Za-z0-9_-]+$'
      description: Identificador opaco, no secuencial y no autorizante.
    CorrelationId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
      description: Identificador seguro sin datos personales.
    Money:
      type: object
      additionalProperties: false
      required: [amountInCents, currency]
      properties:
        amountInCents:
          type: integer
          minimum: 0
          maximum: 999999999999
          description: Importe entero; nunca float.
        currency:
          type: string
          const: COP
      example: { amountInCents: 2500000, currency: COP }
    ProductPublic:
      type: object
      additionalProperties: false
      required: [productId, sku, name, description, imageUrl, unitPrice, available]
      properties:
        productId:
          $ref: '#/components/schemas/OpaqueId'
        sku:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 1000
        imageUrl:
          type: string
          format: uri
          maxLength: 2048
        unitPrice:
          $ref: '#/components/schemas/Money'
        available:
          type: integer
          minimum: 0
          maximum: 999999
          readOnly: true
          description: Proyección pública derivada; no es autoridad para reservar.
    ProductResponse:
      $ref: '#/components/schemas/ProductPublic'
    ProductCollectionResponse:
      type: object
      additionalProperties: false
      required: [items, count]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/ProductPublic' }
        count:
          type: integer
          minimum: 0
          maximum: 100
...
    CanonicalErrorCode:
      type: string
      enum:
        - REQUEST_MALFORMED
        - ORIGIN_FORBIDDEN
        - CHECKOUT_NOT_FOUND_OR_FORBIDDEN
        - PRODUCT_NOT_FOUND
        - FIELD_INVALID
        - OUT_OF_STOCK
        - QUOTE_STALE
        - CHECKOUT_EXPIRED
        - PRECONDITION_FAILED
        - IDEMPOTENCY_CONFLICT
        - PAYMENT_ALREADY_IN_PROGRESS
        - PAYMENT_TOKEN_REJECTED
        - PROVIDER_NOT_SENT
        - PROVIDER_OUTCOME_UNKNOWN
        - WEBHOOK_SIGNATURE_INVALID
        - WEBHOOK_DUPLICATE
        - WEBHOOK_OUT_OF_ORDER
        - STATE_TRANSITION_CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
        - FINAL_STATE_CONFLICT
        - APPROVED_INVENTORY_CONFLICT
        - ENVIRONMENT_MISMATCH
        - PROVIDER_AUTH_OR_CONFIG_INVALID
    ProblemField:
      type: object
      additionalProperties: false
      required: [path, code, reason]
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9_.\[\]-]+$'
        code:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-z0-9_]+$'
        reason:
          type: string
          minLength: 1
          maxLength: 200
          description: Razón allowlisted; nunca contiene el valor rechazado.
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, correlationId, retryable]
      properties:
        type:
          type: string
          format: uri
          maxLength: 2048
        title:
          type: string
          minLength: 1
          maxLength: 120
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
          maxLength: 500
          description: Detalle seguro sin PII, secretos, stack ni respuesta cruda externa.
        instance:
          type: string
          format: uri-reference
          maxLength: 2048
        code:
          $ref: '#/components/schemas/CanonicalErrorCode'
        correlationId:
          $ref: '#/components/schemas/CorrelationId'
        retryAfterSeconds:
          type: integer
          minimum: 1
          maximum: 300
        errors:
          type: array
          maxItems: 50
          items: { $ref: '#/components/schemas/ProblemField' }
        retryable:
          type: boolean
          description: Se refiere a repetir el mismo método y URI; nunca autoriza un POST ambiguo.
        currentState:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[A-Z0-9_]+$'
      description: RFC 9457 exclusivamente para respuestas HTTP 4xx y 5xx.